+++ components/admin/UserDetailModal.tsx
@@ -8,7 +8,7 @@ import { cn } from "@/lib/utils"
 
 function formatDate(date: Date | string | null): string {
   if (!date) return '—'
-  return new Date(date).toLocaleDateString('en-US', { month: 'short', day: 'numeric', year: 'numeric' })
+  return new Date(date).toLocaleDateString('en-US', { month: 'long', day: 'numeric', year: 'numeric' })
 }
 
 type UserDetailModalProps = {
