+++ components/admin/communication/AdminMessagingPanel.tsx
@@ -35,7 +35,7 @@ function formatRelativeTime(date: Date): string {
     if (diffHours < 24) return `${diffHours}h ago`
     if (diffDays === 1) return 'Yesterday'
     if (diffDays < 7) return `${diffDays}d ago`
-    return date.toLocaleDateString('en-US', { month: 'short', day: 'numeric' })
+    return date.toLocaleDateString('en-US', { month: 'long', day: 'numeric' })
 }
 
 type SearchUser = {
