+++ components/dashboard/HomeCalendarView.tsx
@@ -252,7 +252,7 @@ function RequestEventModal({
             <div className="flex items-center gap-3 mt-2 text-sm text-gray-500">
               <span className="flex items-center gap-1">
                 <Calendar className="w-3 h-3" />
-                {eventDate.toLocaleDateString('en-US', { weekday: 'short', month: 'short', day: 'numeric' })}
+                {eventDate.toLocaleDateString('en-US', { weekday: 'long', month: 'long', day: 'numeric' })}
               </span>
               <span className="flex items-center gap-1">
                 <Clock className="w-3 h-3" />
