+++ components/notifications/NotificationList.tsx
@@ -92,7 +92,7 @@ function formatTime(date: Date | string) {
   if (minutes < 60) return `${minutes}m ago`
   if (hours < 24) return `${hours}h ago`
   if (days < 7) return `${days}d ago`
-  return d.toLocaleDateString('en-US', { month: 'short', day: 'numeric' })
+  return d.toLocaleDateString('en-US', { month: 'long', day: 'numeric' })
 }
 
 export function NotificationList({ initialNotifications, compact = false, onNotificationClick }: NotificationListProps) {
